Image sharing website made for a class
An image sharing website (kind of like imgur) made with the Laravel
framework. Made as a class project for a web programming course at my
university.

The application allow registered users to upload images to a public
gallery where other users can comment on them and rate them with an
"upvote" or a "downvote". Voting is implemented using AJAX. The app uses
Sentry 2 for user management and Imagine for manipulating images. The
website's front-end uses Twitter Bootstrap.